+++ src/ISprite.h
@@ -5,6 +5,6 @@ class ISprite
 public:
 
 	virtual void LoadImage(const char* path) = 0;
-	virtual void SetPosition(int x, int y) = 0;
-	virtual void SetSize(int w, int h) = 0;
+	virtual void SetPosition(float x, float y) = 0;
+	virtual void SetSize(float w, float h) = 0;
 };
